pyproject.toml
[project]
name = "psrs-recruitment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "drf-nested-routers>=0.94.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.10",
]
